Added resest password module
