+++ src/main/java/com/org/NoteMakingApp/Dto/UserRequest.java
@@ -1,0 +1,13 @@
+package com.org.NoteMakingApp.Dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserRequest {
+
+	private String oldPassword;
+	private String newPassword;
+
+}
