+++ src/main/java/com/org/NoteMakingApp/service/UserService.java
@@ -1,15 +1,11 @@
 package com.org.NoteMakingApp.service;
 
-import com.org.NoteMakingApp.Dto.LoginRequest;
-import com.org.NoteMakingApp.Dto.LoginResponse;
-import com.org.NoteMakingApp.Dto.UsersDto;
-import com.org.NoteMakingApp.ExceptionHandler.AlreadyExists;
-
-import jakarta.servlet.http.HttpServletRequest;
+import com.org.NoteMakingApp.Dto.UserRequest;
+import com.org.NoteMakingApp.Dto.UserResponse;
 
 public interface UserService {
 
-	public boolean registerUser(UsersDto userdto, String reqUrl) throws AlreadyExists, Exception;
-	
-	public LoginResponse login(LoginRequest loginRequest);
+	public UserResponse userProfile();
+
+	public void resetPassword(UserRequest userRequest);
 }
